nit: Move methods to sessions.go
